Fix: Replace deprecated torch_dtype with dtype
Changed AutoModelForCausalLM.from_pretrained parameter:
- torch_dtype=self.dtype (deprecated)
+ dtype=self.dtype (current)

This removes the deprecation warning during model loading.

+++ mira/core/model_wrapper.py
@@ -110,11 +110,11 @@ class ModelWrapper:
             
             # Load model with proper dtype for GPU memory efficiency
             # Note: For GPU models, this uses float16 (set in self.dtype)
-            # The torch_dtype parameter controls the precision during loading
+            # The dtype parameter controls the precision during loading
             self.model = AutoModelForCausalLM.from_pretrained(
                 load_path,
                 cache_dir=cache_dir,
-                torch_dtype=self.dtype,  # float16 for GPU, float32 for CPU
+                dtype=self.dtype,  # float16 for GPU, float32 for CPU
                 device_map=None,  # Manual device placement with .to()
                 trust_remote_code=True,
                 attn_implementation="eager",  # Enable output_attentions
